refactor: extract infmon-ipc unit tests into _tests.rs companion files
Extract inline #[cfg(test)] test modules from 4 IPC source files into
dedicated companion test files:
- types.rs -> types_tests.rs (2 tests)
- decode.rs -> decode_tests.rs (5 tests)
- control_client.rs -> control_client_tests.rs (1 test)
- stats_client.rs -> stats_client_tests.rs (3 tests)

Closes #107
