+++ src/common/src/ipc/control_client.rs
@@ -74,12 +74,5 @@ impl InFMonControlClient {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn new_stores_path() {
-        let client = InFMonControlClient::new(Path::new("/tmp/test.sock"));
-        assert_eq!(client.path(), Path::new("/tmp/test.sock"));
-    }
-}
+#[path = "control_client_tests.rs"]
+mod control_client_tests;
